Moved characteristics from DataSource to Dataset
