+++ silk-core/src/main/scala/org/silkframework/dataset/EmptySource.scala
@@ -33,6 +33,4 @@ object EmptySource extends DataSource {
   }
 
   override def underlyingTask: Task[DatasetSpec[Dataset]] = PlainTask("empty_dataset", DatasetSpec(EmptyDataset))
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics()
 }
