+++ silk-core/src/main/scala/org/silkframework/dataset/EntityDatasource.scala
@@ -54,6 +54,4 @@ case class EntityDatasource(underlyingTask: Task[DatasetSpec[Dataset]], entities
                             (implicit userContext: UserContext): IndexedSeq[TypedPath] = {
     entitySchema.typedPaths
   }
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics()
 }
