+++ silk-core/src/main/scala/org/silkframework/dataset/rdf/RdfDataset.scala
@@ -1,6 +1,7 @@
 package org.silkframework.dataset.rdf
 
-import org.silkframework.dataset.Dataset
+import org.silkframework.dataset.DataSourceCharacteristics.SupportedPathExpressions
+import org.silkframework.dataset.{DataSourceCharacteristics, Dataset}
 
 trait RdfDataset extends Dataset {
 
@@ -10,5 +11,17 @@ trait RdfDataset extends Dataset {
     * URI of the graph this RDF dataset is referring to, if any.
     */
   def graphOpt: Option[String] = None
+
+  /** Shared characteristics of all RDF Datasets */
+  override final val characteristics: DataSourceCharacteristics = {
+    DataSourceCharacteristics(
+      SupportedPathExpressions(
+        multiHopPaths = true,
+        backwardPaths = true,
+        propertyFilter = true,
+        languageFilter = true
+      )
+    )
+  }
 
 }
