+++ silk-core/src/test/scala/org/silkframework/dataset/MockDataset.scala
@@ -25,6 +25,8 @@ case class MockDataset(@Param(label = "person name", value = "The full name of a
   override def linkSink(implicit userContext: UserContext): LinkSink = DummyLinkSink(writeLinkFn, clearFn)
 
   override def entitySink(implicit userContext: UserContext): EntitySink = DummyEntitySink(writeEntityFn, clearFn)
+
+  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics.attributesOnly
 }
 
 case class DummyDataSource(retrieveFn: (EntitySchema, Option[Int]) => Traversable[Entity],
@@ -49,8 +51,6 @@ case class DummyDataSource(retrieveFn: (EntitySchema, Option[Int]) => Traversabl
                             (implicit userContext: UserContext): Traversable[(String, Double)] = Traversable.empty
 
   override def underlyingTask: Task[DatasetSpec[Dataset]] = EmptySource.underlyingTask
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics()
 }
 
 case class DummyLinkSink(writeLinkFn: (Link, String) => Unit,
