+++ silk-plugins/silk-plugins-csv/src/main/scala/org/silkframework/plugins/dataset/csv/CsvDatasetTrait.scala
@@ -1,5 +1,6 @@
 package org.silkframework.plugins.dataset.csv
 
+import org.silkframework.dataset.DataSourceCharacteristics
 import org.silkframework.plugins.dataset.charset.CharsetUtils
 
 import scala.io.Codec
@@ -52,4 +53,6 @@ trait CsvDatasetTrait {
 
   protected val csvSettings: CsvSettings = CsvSettings(separatorChar, arraySeparatorChar, quoteChar,
     maxCharsPerColumn = Some(maxCharsPerColumn), quoteEscapeChar = quoteEscapeChar, linesToSkip = linesToSkip, codec = codec)
+
+  def characteristics: DataSourceCharacteristics = DataSourceCharacteristics.attributesOnly
 }
