+++ silk-plugins/silk-plugins-csv/src/main/scala/org/silkframework/plugins/dataset/csv/CsvSource.scala
@@ -390,10 +390,6 @@ class CsvSource(file: Resource,
     * @return
     */
   override def underlyingTask: Task[DatasetSpec[Dataset]] = PlainTask(Identifier.fromAllowed(file.name), DatasetSpec(EmptyDataset))   //FIXME CMEM-1352 replace with actual task
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics(
-    SupportedPathExpressions()
-  )
 }
 
 object CsvSource {
