+++ silk-plugins/silk-plugins-json/src/main/scala/org/silkframework/plugins/dataset/json/JsonDataset.scala
@@ -1,5 +1,6 @@
 package org.silkframework.plugins.dataset.json
 
+import org.silkframework.dataset.DataSourceCharacteristics.{SpecialPathInfo, SupportedPathExpressions}
 import org.silkframework.dataset._
 import org.silkframework.plugins.dataset.hierarchical.HierarchicalSink.DEFAULT_MAX_SIZE
 import org.silkframework.runtime.activity.UserContext
@@ -37,4 +38,28 @@ case class JsonDataset(
   override def linkSink(implicit userContext: UserContext): LinkSink = new TableLinkSink(new JsonSink(file, maxDepth = maxDepth))
 
   override def entitySink(implicit userContext: UserContext): EntitySink = new JsonSink(file, jsonTemplate, maxDepth)
+
+  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics(
+    SupportedPathExpressions(
+      multiHopPaths = true,
+      backwardPaths = true,
+      propertyFilter = true,
+      specialPaths = Seq(
+        SpecialPathInfo(JsonDataset.specialPaths.ID, Some("Hash value of the JSON node or value.")),
+        SpecialPathInfo(JsonDataset.specialPaths.TEXT, Some("The string value of a node. This will turn a JSON object into it's string representation.")),
+        SpecialPathInfo(JsonDataset.specialPaths.BACKWARD_PATH, Some("Navigates back to parent object."))
+      )
+    )
+  )
+}
+
+object JsonDataset {
+
+  object specialPaths {
+    final val TEXT = "#text"
+    final val ID = "#id"
+    final val BACKWARD_PATH = "\\.."
+    final val all = Seq(ID, TEXT, BACKWARD_PATH)
+  }
+
 }
