+++ silk-plugins/silk-plugins-json/src/main/scala/org/silkframework/plugins/dataset/json/JsonSource.scala
@@ -216,19 +216,6 @@ case class JsonSource(taskId: Identifier, input: JsValue, basePath: String, uriP
     * @return
     */
   override lazy val underlyingTask: Task[DatasetSpec[Dataset]] = PlainTask(taskId, DatasetSpec(EmptyDataset))     //FIXME CMEM 1352 replace with actual task
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics(
-    SupportedPathExpressions(
-      multiHopPaths = true,
-      backwardPaths = true,
-      propertyFilter = true,
-      specialPaths = Seq(
-        SpecialPathInfo(JsonSource.specialPaths.ID, Some("Hash value of the JSON node or value.")),
-        SpecialPathInfo(JsonSource.specialPaths.TEXT, Some("The string value of a node. This will turn a JSON object into it's string representation.")),
-        SpecialPathInfo(JsonSource.specialPaths.BACKWARD_PATH, Some("Navigates back to parent object."))
-      )
-    )
-  )
 }
 
 object JsonSource {
@@ -241,12 +228,5 @@ object JsonSource {
     } else {
       apply(Identifier.fromAllowed(file.name), JsArray(), basePath, uriPattern)
     }
-  }
-
-  object specialPaths {
-    final val TEXT = "#text"
-    final val ID = "#id"
-    final val BACKWARD_PATH = "\\.."
-    final val all = Seq(ID, TEXT, BACKWARD_PATH)
   }
 }
