+++ silk-plugins/silk-plugins-rdf/src/main/scala/org/silkframework/plugins/dataset/rdf/access/SparqlSource.scala
@@ -182,17 +182,5 @@ class SparqlSource(params: SparqlParams, val sparqlEndpoint: SparqlEndpoint)
         SparqlSamplePathsCollector(sparqlEndpoint, params.graph, restriction, limit).toIndexedSeq
     }
   }
-
-  override def characteristics: DataSourceCharacteristics = SparqlSource.characteristics
 }
 
-object SparqlSource {
-  final val characteristics: DataSourceCharacteristics = DataSourceCharacteristics(
-    SupportedPathExpressions(
-      multiHopPaths = true,
-      backwardPaths = true,
-      propertyFilter = true,
-      languageFilter = true
-    )
-  )
-}
