+++ silk-plugins/silk-plugins-rdf/src/main/scala/org/silkframework/plugins/dataset/rdf/datasets/RdfFileDataset.scala
@@ -191,8 +191,6 @@ case class RdfFileDataset(
     }
 
     private def sparqlSource = new SparqlSource(SparqlParams(graph = graphOpt), endpoint)
-
-    override def characteristics: DataSourceCharacteristics = SparqlSource.characteristics
   }
 
   override def tripleSink(implicit userContext: UserContext): TripleSink = new FormattedEntitySink(file, formatter)
