+++ silk-plugins/silk-plugins-xml/src/main/scala/org/silkframework/plugins/dataset/xml/XmlDataset.scala
@@ -1,5 +1,6 @@
 package org.silkframework.plugins.dataset.xml
 
+import org.silkframework.dataset.DataSourceCharacteristics.{SpecialPathInfo, SupportedPathExpressions}
 import org.silkframework.dataset._
 import org.silkframework.dataset.bulk.BulkResourceBasedDataset
 import org.silkframework.runtime.activity.UserContext
@@ -50,6 +51,8 @@ case class XmlDataset( @Param("The XML file. This may also be a zip archive of m
 
   override def entitySink(implicit userContext: UserContext): EntitySink = new XmlSink(file, outputTemplate.str)
 
+  override def characteristics: DataSourceCharacteristics = XmlDataset.characteristics
+
   /**
     * Validates the output template parameter
     */
@@ -81,5 +84,33 @@ case class XmlDataset( @Param("The XML file. This may also be a zip archive of m
           "instruction of the form <?Entity?>!")
     }
   }
+
+}
+
+object XmlDataset {
+
+  object SpecialXmlPaths {
+    final val ID = "#id"
+    final val TAG = "#tag"
+    final val TEXT = "#text"
+    final val ALL_CHILDREN = "*"
+    final val ALL_CHILDREN_RECURSIVE = "**"
+    final val BACKWARD_PATH = "\\.."
+  }
+  import SpecialXmlPaths._
+  final val characteristics = DataSourceCharacteristics(
+    SupportedPathExpressions(
+      multiHopPaths = true,
+      propertyFilter = true,
+      specialPaths = Seq(
+        SpecialPathInfo(BACKWARD_PATH, Some("Navigate to parent element.")),
+        SpecialPathInfo(ID, Some("A document-wide unique ID of the entity.")),
+        SpecialPathInfo(TAG, Some("The element tag of the entity.")),
+        SpecialPathInfo(TEXT, Some("The concatenated text inside an element.")),
+        SpecialPathInfo(ALL_CHILDREN, Some("Selects all direct children of the entity.")),
+        SpecialPathInfo(ALL_CHILDREN_RECURSIVE, Some("Selects all children nested below the entity at any depth."))
+      )
+    )
+  )
 
 }
