+++ silk-plugins/silk-plugins-xml/src/main/scala/org/silkframework/plugins/dataset/xml/XmlSourceInMemory.scala
@@ -109,8 +109,6 @@ class XmlSourceInMemory(file: Resource, basePath: String, uriPattern: String) ex
     * @return
     */
   override def underlyingTask: Task[DatasetSpec[Dataset]] = PlainTask(Identifier.fromAllowed(file.name), DatasetSpec(EmptyDataset))   //FIXME CMEM-1352 replace with actual task
-
-  override def characteristics: DataSourceCharacteristics = XmlSourceTrait.characteristics
 }
