+++ silk-plugins/silk-plugins-xml/src/main/scala/org/silkframework/plugins/dataset/xml/XmlSourceStreaming.scala
@@ -457,6 +457,4 @@ class XmlSourceStreaming(file: Resource, basePath: String, uriPattern: String) e
     * @return
     */
   override def underlyingTask: Task[DatasetSpec[Dataset]] = PlainTask(Identifier.fromAllowed(file.name), DatasetSpec(EmptyDataset))   //FIXME CMEM-1352 replace with actual task
-
-  override def characteristics: DataSourceCharacteristics = XmlSourceTrait.characteristics
 }
