+++ silk-rules/src/main/scala/org/silkframework/rule/TransformedDataSource.scala
@@ -1,14 +1,14 @@
 package org.silkframework.rule
 
 import org.silkframework.config.Task
-import org.silkframework.dataset.{DataSource, DataSourceCharacteristics, Dataset, DatasetSpec, SupportedPathExpressions}
+import org.silkframework.dataset.{DataSource, Dataset, DatasetSpec}
 import org.silkframework.entity.metadata.GenericExecutionFailure
 import org.silkframework.entity.paths.TypedPath
 import org.silkframework.entity.{Entity, EntitySchema}
 import org.silkframework.execution.EntityHolder
 import org.silkframework.execution.local.{EmptyEntityTable, GenericEntityTable}
-import org.silkframework.runtime.activity.UserContext
 import org.silkframework.failures.FailureClass
+import org.silkframework.runtime.activity.UserContext
 import org.silkframework.util.Uri
 
 import scala.util.{Failure, Success, Try}
@@ -128,8 +128,4 @@ class TransformedDataSource(source: DataSource, inputSchema: EntitySchema, trans
     * @return
     */
   override def underlyingTask: Task[DatasetSpec[Dataset]] = source.underlyingTask
-
-  override def characteristics: DataSourceCharacteristics = DataSourceCharacteristics(SupportedPathExpressions(
-    backwardPaths = true
-  ))
 }
